requirements.txt
asgiref==3.4.1
black==21.10b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.10
django-cors-headers==3.7.0
django-environ==0.8.1
django-filter==2.0.0
djangorestframework==3.10.0
drf-yasg==1.17.0
idna==3.3
importlib-metadata==4.8.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.2
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2021.3
regex==2021.11.2
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
